nonlinear-covariate-gwas/requirements.txt
wheel>=0.36
absl-py>=0.12
ml_collections>=0.1
numpy>=1.19
pandas>=1.1
tensorflow>=2.4.1
tensorflow-probability>=0.12
xgboost>=1.4
